[armv8vp] Memory architecture + first instruction steps
Former-commit-id: c1cd1c31ae204bd88adf8f632c6cd83cad6e7bd4

+++ unisim_simulators/cxx/vle4fuzr/emu.hh
@@ -41,7 +41,7 @@ struct Processor
     {
       using is_transparent = void;
       bool operator() (Page const& a, Page const& b) const { return a > b; }
-      bool operator() (Page const& a, unsigned b) const { return a.base > b; }
+      bool operator() (Page const& a, uint64_t b) const { return a.base > b; }
     };
     Page( Page&& page )
       : base(page.base), last(page.last), data(page.data)
